requirements.txt
Django==3.0.8
django-cors-headers==3.2.1
django-filter==2.2.0
django-model-utils==4.0.0
django-rest-knox>=4.0.0,<5.0.0
djangorestframework==3.11.0
rest_condition==1.0.3
rest-social-auth==4.2.0
social-auth-core>=3.0,<4.0
social-auth-app-django==3.1.0
geopy==1.21.0
psycopg2-binary==2.8.4
